real_time_data_streaming/requirements.txt
aws-cdk.core
aws-cdk.aws-kinesis
aws-cdk.aws-lambda
aws-cdk.aws-iam
boto3
